docker_scripts/finetune_trocr/export_dataset/requirements.txt
supabase
pillow
pillow-avif-plugin
pandas
tqdm
scikit-learn
requests
python-dotenv
